+++ src/config.php
@@ -33,7 +33,9 @@ return array(
     // Admin Bar PRO
     'displayWidgetLabels' => false,
     'widgetEnabledBlitz' => false,
+    'widgetEnabledCraftAuthors' => false,
     'widgetEnabledCraftNewEntry' => false,
+    'widgetEnabledCraftPublished' => false,
     'widgetEnabledCraftSites' => false,
     'widgetEnabledGuide' => false,
     'widgetEnabledSeomatic' => false,